Traversal now lets Assembly know what Bio::Network::ProteinNet it is associated with.
Assembly now groups based on graph topology, which is the connections between the orginal components, regardless of the templates used to connect the components in the solution. It tries to approximate what solutions contain the same components connected to each other, even if the used templates might make the structure quite different.