Fixed a bug in Transform::loadstr that was not removing whitespace in the input
DomIO is basically working now, to read a DOM file into a stream of CofM objects.
Now, need to get this into an Assembly object.

CofM now uses a 'tainted' field to note when it is no longer on the identity transform.

Now need a Assembly::rmsd(Assembly), based on centres-of-mass of Dom's