Keeping track of node and template covering during traversal. These are no passed back to callback when a partial solution reached.
The 'test' callback function also now gets the name of the template, allowing one to look it up $graph->get_edge_attribute_by_id($template_id);

Also, tracking of available alternatives between two nodes is now done internal to Traversal. This no longer has to be managed by the 'test' callback. As mentioned, it just gets the name of the current template and just needs to return whether that one template can/should be used or not.