bin/sbgstamp replaces do_stamp (using SBG:: tools)
bin/apply applies the Transform in a *.trans file to all Domains in a list of DOM files
  IDs are appended with a counter, counting the number of transformations applied. 
  This is also reflected in the output file names of the transformed files/domains

bin/superpose creates a Transform in a *.trans file, given two DOM files. The transform is the transformation required to superpose the first domain onto the second. Each file should contain just one domain (only the first domain of each is read)

t/STAMP.t tests the exosome case and verifies that the SBG:: tools return the same results as the command line STAMP tools.