Re-read the logic on the rotations. Finally think it's figured out. However :
Still trying to fix GeometricHash and rotations.
Don't know if it's broken or if the contrived test harness is broken.
Can't seem to hash an object correctly after rotating it.

Added component mapping to hash.
Necessary when the points in space have an actual identity that needs to be retained.
Labels are optional. Works without them as well. Then, matching is simply geometric.