Set to version 1.00
All tests passing, both on 32bit Ubuntu and on 64bit CentOS

Added Versionable role in order for Storable objects to save the version of the module with which they were created.
Applies to Complex and Network for now.

Simplified config.ini (to be removed)

Cached objects now default to a subdir of $CACHEDIR

STAMP now just looks for its binary in the PATH. So this should be something like PATH=$PATH:$STAMPDIR/../bin/linux