Fixed RMSD::superpose. Order of operations was wrong. In STAMP, the rotation matrix is based on the rotation when both domains are supeposed the origin. But the translation STAMP uses is based on the found rotation being applied to the domain at its original location, rotating about the origin from there. Then the translation required to translate that rotated domain onto the reference domain is calculated. This is all consistent with STAMP now.
This also verfied that RMSD::rmsd works accurately. And verified superposition visually as well.

RMSD.t wrongly thought it was succeeding before. This was only because the test data was poorly chosen, providing ambiguous results. 1TIMA and 1TIMB shouldn't be used for testing superposition since A->B and B->A are just related by 2-fold rotational symmetry. Furthermore, using Domain::Atoms, one must make sure that the two domains have the same number of atoms, and these should be very homologous in order for the superposition to succeed and be comparable to what STAMP would do. Now using 2br2 B 10 _ to B 240 _ and D 10 _ to D 240 _