lib/SBG/Traversal.pm
Traversal uses Heap::Priority as a heuristic
  Nodes prioritized by Sc of linker_superpositino
  Edges prioritized by score of next alternative interaction
    Interaction scored by frac_conserved * nres along interface (avg)

Alternatives on an edge now stored in a list in the Traversal object.
  No longer requires looking up/sorting every time.
  Index just walks throught the appropriate list for a given edge

lib/SBG/Complex.pm

domains() now accepts an argument $keys that defines the keys of the domain  objects to be returned.

merge() accepts an option keys=>$keys to combine in the domain returned

add_interaction() returns the Sc score of the linker_superposition when successful, otherwise returns nothing

rmsd() now also uses the Superposition cache

lib/SBG/CA/Assembler.pm

score() returns the score of the edge alternative with the given ID