Using Storable::dclone rather than Clone::clone now
No longer considering subs::parallel
It is built on ithreads, which uses the CLONE method to duplicate local data to the newly spawned thread. As many of our objects contain PDL instances, this may well lead to problems, as both PDL objects will continue to refer to the same underlying XS memory. PDL does not appear to be thread-safe.