Graph::Traversal::GreedyEdges
using Storable::dclone rather than Clone::clone
Together with our modified PDL::IO::Storable, this allows us to get the proper cloning semantics (required for the recursive backtracking algorithm)

Fixed bug where we were not checking the stopping condition (having tried every interaction in the list)

No longer storing intermediate state in the GreedyEdges traverser, as the state is a function of the depth in the call. As such, it is passed during recursive function calling, which automatically gets cleaned up when it falls out of scope, obviating the need for manual cleanup of a hash saved in the GreedEdges traverser.

No longer storing subcomplex models in the SBG::Network. The use of refvertex=>0 seems to prevent this. We need refvertex=>0 in SBG::Network due a bug in Graph::AdjacencyMap (described in documentation for SBG::Network).

Only calling 'solution' when more than 2 components in a model.

Now using simple 'goto' for optimization of tail recursion. The Sub::Call::Recur seemed to be having undefinable problems. Tail recursion is not optimized away during a debugging session, however (when DB::sub is defined).

No longer using SBG::U::List::reorder to a bug whereby duplicate entries are squashed. Just using builtin 'sort' now, and sorting on the field of SBG::Interaction given in Greedy::Edges->sorter