res_mapping::aln2locations() takes a Bio::SimpleAlign and extracts a bijection between the sequence coordinates of each of the two aligned sequences.
res_mapping::query() takes an ArrayRef rather than $start, $end now. This should contain all the positions in between that also need to be fetched.
Rows are returned according to the order in sequence, not in the structure, i.e. the residue IDs returned are not necessarily in order, but they do correspond to the order of the sequence given, as long as the sequence is ordered.